+++ bindings/nostr-sdk-ffi/src/logger.rs
@@ -2,6 +2,7 @@
 // Copyright (c) 2023-2024 Rust Nostr Developers
 // Distributed under the MIT software license
 
+use nostr_ffi::NostrLibrary;
 use tracing::Level;
 #[cfg(any(target_os = "android", target_os = "ios"))]
 use tracing_subscriber::filter::Targets;
@@ -44,7 +45,15 @@ pub fn init_logger(level: LogLevel) {
         .with_max_level(level)
         .finish();
     match tracing::subscriber::set_global_default(subscriber) {
-        Ok(_) => tracing::info!("Desktop logger initialized"),
+        Ok(_) => {
+            tracing::info!("Desktop logger initialized");
+
+            // Log git hash (defined at compile time)
+            match NostrLibrary::new().git_hash_version() {
+                Some(hash) => tracing::info!("Git hash: {hash}"),
+                None => tracing::warn!("Git hash not defined!"),
+            };
+        }
         Err(e) => eprintln!("Impossible to init desktop logger: {e}"),
     }
 }
@@ -69,7 +78,15 @@ pub fn init_logger(level: LogLevel) {
         .try_init();
 
     match res {
-        Ok(_) => tracing::info!("Mobile logger initialized"),
+        Ok(_) => {
+            tracing::info!("Mobile logger initialized");
+
+            // Log git hash (defined at compile time)
+            match NostrLibrary::new().git_hash_version() {
+                Some(hash) => tracing::info!("Git hash: {hash}"),
+                None => tracing::warn!("Git hash not defined!"),
+            };
+        }
         Err(e) => eprintln!("Impossible to init mobile logger: {e}"),
     }
 }